Add printing function

+++ src/main/java/com/concrete/concrete_plant_management/client/ClientService.java
@@ -1,5 +1,7 @@
 package com.concrete.concrete_plant_management.client;
 
+import com.concrete.concrete_plant_management.order.OrderService;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -8,14 +10,20 @@ import java.util.List;
 public class ClientService {
 
     private final ClientRepository repository;
+    private final OrderService orderService;
 
-    public ClientService(final ClientRepository repository) {
+    public ClientService(final ClientRepository repository,
+                         final OrderService orderService) {
         this.repository = repository;
+        this.orderService = orderService;
     }
 
     Client saveClient(Client toSave) {
-        ClientSaveModel clientSaveModel = new ClientSaveModel();
-        Client result = clientSaveModel.clientModelValidating(toSave);
+        if (repository.existsClientByNip(toSave.getNip())){
+            return null;
+        }
+        ClientDataValidation clientDataValidation = new ClientDataValidation();
+        Client result = clientDataValidation.clientModelValidating(toSave);
         return repository.save(result);
     }
 
@@ -27,19 +35,26 @@ public class ClientService {
         return repository.findAll();
     }
 
-    Client updateClient(final Client toUpdate) {
-        ClientSaveModel clientSaveModel = new ClientSaveModel();
-        Client result = clientSaveModel.clientModelValidating(toUpdate);
-        repository.findById(toUpdate.getId()).ifPresent(entity -> repository.save(result));
+    Client updateClient(final int id, final Client toUpdate) {
+        if (!repository.existsById(id)) {
+            return null;
+        }
+        ClientDataValidation clientDataValidation = new ClientDataValidation();
+        Client result = clientDataValidation.clientModelValidating(toUpdate);
+        repository.save(toUpdate);
         return result;
     }
 
     public boolean deleteClient(final int id) {
-        if (repository.existsById(id)){
+        if (repository.existsById(id) && !orderService.existsOrderByClientId(id)){
             repository.deleteById(id);
             return true;
         }else{
             return false;
         }
+    }
+
+    public List<Client> getAllClients(final Sort sort) {
+        return repository.findAll(sort);
     }
 }
