+++ src/main/java/com/concrete/concrete_plant_management/vehicle/VehicleService.java
@@ -1,5 +1,7 @@
 package com.concrete.concrete_plant_management.vehicle;
 
+import com.concrete.concrete_plant_management.order_batch.OrderBatchService;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -8,22 +10,28 @@ import java.util.List;
 public class VehicleService {
 
     private final VehicleRepository repository;
+    private final OrderBatchService orderBatchService;
 
-    public VehicleService(final VehicleRepository repository) {
+    public VehicleService(final VehicleRepository repository,
+                          final OrderBatchService orderBatchService) {
         this.repository = repository;
+        this.orderBatchService = orderBatchService;
     }
 
     Vehicle saveVehicle(final Vehicle toSave) {
-        toSave.setRegNo(toSave.getRegNo().toUpperCase());
+        if (repository.existsByRegNo(toSave.getRegNo())){
+            return null;
+        }
+        new VehicleDataValidation().vehicleDataValidation(toSave);
         return repository.save(toSave);
     }
 
     Vehicle vehicleUpdate(final int id, final Vehicle toUpdate) {
-        if (repository.existsById(id)) {
-            return repository.save(toUpdate);
-        } else {
+        if (!repository.existsById(id)) {
             return null;
         }
+        new VehicleDataValidation().vehicleDataValidation(toUpdate);
+        return repository.save(toUpdate);
     }
 
     Vehicle readVehicle(final int id) {
@@ -34,8 +42,12 @@ public class VehicleService {
         return repository.findAll();
     }
 
+    public List<Vehicle> readAllVehicles(final Sort sort) {
+        return repository.findAll(sort);
+    }
+
     boolean deleteVehicle(final int id) {
-        if (repository.existsById(id)) {
+        if (repository.existsById(id) && !orderBatchService.existsOrderBatchByVehicleId(id)) {
             repository.deleteById(id);
             return true;
         } else {
